# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Assinatura Streaming",
        "version": "1.0.0"
    },
    "paths": {
        "/assinaturas": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Lista todas as assinaturas",
                "operationId": "4ea2f37654244a965c9277c6bcf47fe8",
                "responses": {
                    "200": {
                        "description": "Lista de assinaturas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Cria uma nova assinatura",
                "operationId": "a64e9f48e4d7f93e54e4d0e5f10ec714",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "price_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price_id": {
                                        "type": "string",
                                        "example": "price_12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assinatura criada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar assinatura"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assinaturas/{user}": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Exibe uma assinatura",
                "operationId": "3016f668106da3eaae7c0b6a3010e40c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da assinatura"
                    },
                    "404": {
                        "description": "Assinatura não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assinaturas/cancelar": {
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Cancela uma assinatura",
                "operationId": "eaa5e386d06fbad793e97fd6a39d1e6a",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assinatura cancelada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao cancelar assinatura"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assinaturas/reativar": {
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Reativa uma assinatura em período de carência",
                "operationId": "5fa9df2f24ef4d47368d392d3d911918",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assinatura reativada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao reativar assinatura"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assinaturas/status": {
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Verifica o status da assinatura do usuário autenticado",
                "operationId": "32528e5052152976bbc629f884ce1361",
                "responses": {
                    "200": {
                        "description": "Status da assinatura"
                    },
                    "404": {
                        "description": "Sem assinatura ativa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}